Convert character to unsigned before hashing.
Fix #881.

+++ c/object.c
@@ -138,7 +138,7 @@ static uint32_t hashString(const char* key, int length) {
   uint32_t hash = 2166136261u;
 
   for (int i = 0; i < length; i++) {
-    hash ^= key[i];
+    hash ^= (uint8_t)key[i];
     hash *= 16777619;
   }
